fix up tests
