+++ decompiler/config/all-types.gc
@@ -502,12 +502,16 @@
     (sky-draw 3)
     (bucket-4 4) ;; ocean
     (tfrag-tex0 5)
+    (tfrag-0 6)
+    (tfrag-near-0 7)
     ;; merc0 10
     ;; generic0 11
     (bucket-10 10)
     (bucket-11 11)
 
     (tfrag-tex1 12)
+    (tfrag-1 13)
+    (tfrag-near-1 14)
     ;; merc1 17
     ;; generic1 18
     (bucket-17 17)
@@ -519,10 +523,20 @@
 
     (generic-foreground 30) ;; ?
     (alpha-tex0 31)
-    (sky-tex0 32)
+    (tfrag-trans-0 32) ;; also sky blend
+    (tfrag-trans-near-0 33)
+    (tfrag-dirt-0 34)
+    (tfrag-dirt-near-0 35)
+    (tfrag-ice-0 36)
+    (tfrag-ice-near-0 37)
 
     (alpha-tex1 38)
-    (sky-tex1 39)
+    (tfrag-trans-1 39) ;; also sky blend
+    (tfrag-trans-near-1 40)
+    (tfrag-dirt-1 41)
+    (tfrag-dirt-near-1 42)
+    (tfrag-ice-1 43)
+    (tfrag-ice-near-1 44)
 
     (bucket-45 45)
     (bucket-46 46)
