+++ game/CMakeLists.txt
@@ -95,6 +95,7 @@ set(RUNTIME_SOURCE
         graphics/opengl_renderer/SkyRenderer.cpp
         graphics/opengl_renderer/SpriteRenderer.cpp
         graphics/opengl_renderer/TextureUploadHandler.cpp
+        graphics/opengl_renderer/tfrag/program6_cpu.cpp
         graphics/opengl_renderer/tfrag/tfrag_unpack.cpp
         graphics/opengl_renderer/tfrag/TFragment.cpp
         graphics/texture/TextureConverter.cpp
