+++ game/graphics/opengl_renderer/BucketRenderer.h
@@ -21,10 +21,10 @@ enum class BucketId {
   SHRUB_TEX_LEVEL0 = 19,
   SHRUB_TEX_LEVEL1 = 25,
   ALPHA_TEX_LEVEL0 = 31,
-  SKY_BLEND_LEVEL0 = 32,
+  TFRAG_TRANS0_AND_SKY_BLEND_LEVEL0 = 32,
   TFRAG_DIRT_LEVEL0 = 34,
   ALPHA_TEX_LEVEL1 = 38,
-  SKY_BLEND_LEVEL1 = 39,
+  TFRAG_TRANS1_AND_SKY_BLEND_LEVEL1 = 39,
   TFRAG_DIRT_LEVEL1 = 41,
   PRIS_TEX_LEVEL0 = 48,
   PRIS_TEX_LEVEL1 = 51,
