+++ goal_src/engine/dma/dma-h.gc
@@ -239,7 +239,7 @@
 
     (generic-foreground 30) ;; ?
     (alpha-tex0 31)
-    (sky-tex0 32) ;; also tfrag trans?
+    (tfrag-trans-0 32) ;; also sky blend
     (tfrag-trans-near-0 33)
     (tfrag-dirt-0 34)
     (tfrag-dirt-near-0 35)
@@ -247,7 +247,7 @@
     (tfrag-ice-near-0 37)
 
     (alpha-tex1 38)
-    (sky-tex1 39) ;; also tfrag trans?
+    (tfrag-trans-1 39) ;; also sky blend
     (tfrag-trans-near-1 40)
     (tfrag-dirt-1 41)
     (tfrag-dirt-near-1 42)
