+++ goal_src/engine/gfx/tfrag/tfrag-h.gc
@@ -42,7 +42,7 @@
    (dma-level-0       uint32               :offset  32)
    (dma-base          uint32               :offset  36)
    (dma-level-1       uint32               :offset  40)
-   (dma-qwc           uint32               4           :offset-assert 44)
+   (dma-qwc           uint8                4           :offset-assert 44)
    (shader            (inline-array adgif-shader)               :offset  48)
    (num-shaders       uint8                :offset  52)
    (num-base-colors   uint8                :offset  53)
