+++ goal_src/engine/gfx/tfrag/tfrag-methods.gc
@@ -267,13 +267,13 @@
         (the-as
          bucket-id
          (if (zero? (-> (scratchpad-object terrain-context) bsp lev-index))
-          32
-          39
+          (bucket-id tfrag-trans-0)
+          (bucket-id tfrag-trans-1)
           )
          )
         s3-0
         (the-as (pointer dma-tag) a3-3)
-        ) |#
+        )|#
        )
       )
      #| TODO
